Addded Errors handlerer for int max value

+++ philo/includes/philo.h
@@ -6,7 +6,7 @@
 /*   By: muayna <muayna@student.42istanbul.com.t    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/04/26 13:30:16 by muayna            #+#    #+#             */
-/*   Updated: 2026/04/26 16:49:13 by muayna           ###   ########.fr       */
+/*   Updated: 2026/04/26 21:46:36 by muayna           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,11 +27,11 @@ Others value is requaried, Must not empty.
 
 typedef struct s_args
 {
-    int number_of_phil;
-    int time_to_die;
-    int time_to_eat;
-    int time_to_sleep;
-    int eat_count;
+    long number_of_phil;
+    long time_to_die;
+    long time_to_eat;
+    long time_to_sleep;
+    long eat_count;
 }t_args;
